fix linter errors, rename folder
