+++ src/floating-select-infinite/floating-select-infinite-menu-option.component.jsx
