+++ src/floating-select-infinite/floating-select-infinite.scss
