+++ src/index.js
@@ -1,4 +1,4 @@
 export * from './floating-select/index';
 export * from './floating-select-portal/index';
-export * from './floating-select-windowed/index';
+export * from './floating-select-infinite/index';
 export styles from './floating-select/styles';
